refactor: Move contenttracker in as part of content
