+++ common/src/main/java/com/wynntils/features/combat/ContentTrackerFeature.java
@@ -10,7 +10,7 @@ import com.wynntils.core.config.Config;
 import com.wynntils.core.config.ConfigCategory;
 import com.wynntils.core.config.RegisterConfig;
 import com.wynntils.core.features.Feature;
-import com.wynntils.models.contenttracker.event.ContentTrackerUpdatedEvent;
+import com.wynntils.models.content.event.ContentTrackerUpdatedEvent;
 import com.wynntils.utils.mc.McUtils;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
